backend/requirements.txt
django>=5.1
channels>=4.0
channels-redis>=4.2
daphne>=4.1
psycopg[binary]>=3.2
python-dotenv>=1.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=12.0

